+++ src/SamplePlayer.cpp
@@ -1,6 +1,6 @@
 #include <SamplePlayer.h>
 
-SamplePlayer::SamplePlayer(AudioPlaySdWav *players[], const size_t n_players)
+SamplePlayer::SamplePlayer(AudioPlaySdWav *players[], const uint8_t n_players)
     : _players(players), _n_players(n_players)
 {
 
@@ -14,9 +14,9 @@ const char *SamplePlayer::getSampleName(int i)
 void SamplePlayer::playSample(int i)
 {
     const char* filename = getSampleName(i);
-    size_t selected = 0;
+    uint8_t selected = 0;
     uint32_t max_pos = 0;
-    for(size_t i = 0 ; i < _n_players ; i++) {
+    for(uint8_t i = 0 ; i < _n_players ; i++) {
         if(!_players[i]->isPlaying()) {
             selected = i;
             break;
